Resolves issue #314 - OPML export: Choose export folder
