+++ app/src/main/java/de/danoeh/antennapod/fragment/preferences/StoragePreferencesFragment.java
@@ -4,6 +4,7 @@ import android.Manifest;
 import android.annotation.SuppressLint;
 import android.app.Activity;
 import android.app.ProgressDialog;
+import android.content.ActivityNotFoundException;
 import android.content.Context;
 import android.content.Intent;
 import android.content.pm.PackageManager;
@@ -13,13 +14,16 @@ import android.os.Build;
 import android.os.Bundle;
 import android.support.v4.app.ActivityCompat;
 import android.support.v4.content.FileProvider;
+import android.support.v4.provider.DocumentFile;
 import android.support.v7.app.AlertDialog;
 import android.support.v7.preference.PreferenceFragmentCompat;
 import android.util.Log;
+
 import de.danoeh.antennapod.R;
 import de.danoeh.antennapod.activity.DirectoryChooserActivity;
 import de.danoeh.antennapod.activity.ImportExportActivity;
 import de.danoeh.antennapod.activity.OpmlImportFromPathActivity;
+import de.danoeh.antennapod.asynctask.DocumentFileExportWorker;
 import de.danoeh.antennapod.asynctask.ExportWorker;
 import de.danoeh.antennapod.core.export.ExportWriter;
 import de.danoeh.antennapod.core.export.html.HtmlWriter;
@@ -45,6 +49,7 @@ public class StoragePreferencesFragment extends PreferenceFragmentCompat {
             Manifest.permission.READ_EXTERNAL_STORAGE,
             Manifest.permission.WRITE_EXTERNAL_STORAGE };
     private static final int PERMISSION_REQUEST_EXTERNAL_STORAGE = 41;
+    private static final int CHOOSE_OPML_EXPORT_PATH = 1;
     private Disposable disposable;
 
     @Override
@@ -59,6 +64,12 @@ public class StoragePreferencesFragment extends PreferenceFragmentCompat {
         setDataFolderText();
     }
 
+    @Override
+    public void onStop() {
+        super.onStop();
+        unsubscribeExportSubscription();
+    }
+
     private void setupStorageScreen() {
         final Activity activity = getActivity();
 
@@ -69,7 +80,11 @@ public class StoragePreferencesFragment extends PreferenceFragmentCompat {
                 }
         );
         findPreference(PREF_OPML_EXPORT).setOnPreferenceClickListener(
-                preference -> export(new OpmlWriter()));
+                preference -> {
+                    openExportPathPicker();
+                    return true;
+                }
+        );
         findPreference(PREF_HTML_EXPORT).setOnPreferenceClickListener(
                 preference -> export(new HtmlWriter()));
         findPreference(PREF_OPML_IMPORT).setOnPreferenceClickListener(
@@ -129,6 +144,10 @@ public class StoragePreferencesFragment extends PreferenceFragmentCompat {
     }
 
     private boolean export(ExportWriter exportWriter) {
+        return export(exportWriter, null);
+    }
+
+    private boolean export(ExportWriter exportWriter, final Uri uri) {
         Context context = getActivity();
         final ProgressDialog progressDialog = new ProgressDialog(context);
         progressDialog.setMessage(context.getString(R.string.exporting_label));
@@ -136,38 +155,71 @@ public class StoragePreferencesFragment extends PreferenceFragmentCompat {
         progressDialog.show();
         final AlertDialog.Builder alert = new AlertDialog.Builder(context)
                 .setNeutralButton(android.R.string.ok, (dialog, which) -> dialog.dismiss());
-        Observable<File> observable = new ExportWorker(exportWriter).exportObservable();
-        disposable = observable.subscribeOn(Schedulers.io())
-                .observeOn(AndroidSchedulers.mainThread())
-                .subscribe(output -> {
-                    alert.setTitle(R.string.export_success_title);
-                    String message = context.getString(R.string.export_success_sum, output.toString());
-                    alert.setMessage(message);
-                    alert.setPositiveButton(R.string.send_label, (dialog, which) -> {
-                        Uri fileUri = FileProvider.getUriForFile(context.getApplicationContext(),
-                                context.getString(R.string.provider_authority), output);
-                        Intent sendIntent = new Intent(Intent.ACTION_SEND);
-                        sendIntent.putExtra(Intent.EXTRA_SUBJECT,
-                                context.getResources().getText(R.string.opml_export_label));
-                        sendIntent.putExtra(Intent.EXTRA_STREAM, fileUri);
-                        sendIntent.setType("text/plain");
-                        sendIntent.addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION);
-                        if (Build.VERSION.SDK_INT <= Build.VERSION_CODES.KITKAT) {
-                            List<ResolveInfo> resInfoList = context.getPackageManager().queryIntentActivities(sendIntent, PackageManager.MATCH_DEFAULT_ONLY);
-                            for (ResolveInfo resolveInfo : resInfoList) {
-                                String packageName = resolveInfo.activityInfo.packageName;
-                                context.grantUriPermission(packageName, fileUri, Intent.FLAG_GRANT_READ_URI_PERMISSION);
+        if (uri == null) {
+            Observable<File> observable = new ExportWorker(exportWriter).exportObservable();
+            disposable = observable.subscribeOn(Schedulers.io())
+                    .observeOn(AndroidSchedulers.mainThread())
+                    .subscribe(output -> {
+                        alert.setTitle(R.string.export_success_title);
+                        String message = context.getString(R.string.export_success_sum, output.toString());
+                        alert.setMessage(message);
+                        alert.setPositiveButton(R.string.send_label, (dialog, which) -> {
+                            Uri fileUri = FileProvider.getUriForFile(context.getApplicationContext(),
+                                    context.getString(R.string.provider_authority), output);
+                            Intent sendIntent = new Intent(Intent.ACTION_SEND);
+                            sendIntent.putExtra(Intent.EXTRA_SUBJECT,
+                                    context.getResources().getText(R.string.opml_export_label));
+                            sendIntent.putExtra(Intent.EXTRA_STREAM, fileUri);
+                            sendIntent.setType("text/plain");
+                            sendIntent.addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION);
+                            if (Build.VERSION.SDK_INT <= Build.VERSION_CODES.KITKAT) {
+                                List<ResolveInfo> resInfoList = context.getPackageManager().queryIntentActivities(sendIntent, PackageManager.MATCH_DEFAULT_ONLY);
+                                for (ResolveInfo resolveInfo : resInfoList) {
+                                    String packageName = resolveInfo.activityInfo.packageName;
+                                    context.grantUriPermission(packageName, fileUri, Intent.FLAG_GRANT_READ_URI_PERMISSION);
+                                }
                             }
-                        }
-                        context.startActivity(Intent.createChooser(sendIntent,
-                                context.getResources().getText(R.string.send_label)));
-                    });
-                    alert.create().show();
-                }, error -> {
-                    alert.setTitle(R.string.export_error_label);
-                    alert.setMessage(error.getMessage());
-                    alert.show();
-                }, progressDialog::dismiss);
+                            context.startActivity(Intent.createChooser(sendIntent,
+                                    context.getResources().getText(R.string.send_label)));
+                        });
+                        alert.create().show();
+                    }, error -> {
+                        alert.setTitle(R.string.export_error_label);
+                        alert.setMessage(error.getMessage());
+                        alert.show();
+                    }, progressDialog::dismiss);
+        } else {
+            Observable<DocumentFile> observable = new DocumentFileExportWorker(exportWriter, context, uri).exportObservable();
+            disposable = observable.subscribeOn(Schedulers.io())
+                    .observeOn(AndroidSchedulers.mainThread())
+                    .subscribe(output -> {
+                        alert.setTitle(R.string.export_success_title);
+                        String message = context.getString(R.string.export_success_sum, output.getUri());
+                        alert.setMessage(message);
+                        alert.setPositiveButton(R.string.send_label, (dialog, which) -> {
+                            Intent sendIntent = new Intent(Intent.ACTION_SEND);
+                            sendIntent.putExtra(Intent.EXTRA_SUBJECT,
+                                    context.getResources().getText(R.string.opml_export_label));
+                            sendIntent.putExtra(Intent.EXTRA_STREAM, output.getUri());
+                            sendIntent.setType("text/plain");
+                            sendIntent.addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION);
+                            if (Build.VERSION.SDK_INT <= Build.VERSION_CODES.KITKAT) {
+                                List<ResolveInfo> resInfoList = context.getPackageManager().queryIntentActivities(sendIntent, PackageManager.MATCH_DEFAULT_ONLY);
+                                for (ResolveInfo resolveInfo : resInfoList) {
+                                    String packageName = resolveInfo.activityInfo.packageName;
+                                    context.grantUriPermission(packageName, output.getUri(), Intent.FLAG_GRANT_READ_URI_PERMISSION);
+                                }
+                            }
+                            context.startActivity(Intent.createChooser(sendIntent,
+                                    context.getResources().getText(R.string.send_label)));
+                        });
+                        alert.create().show();
+                    }, error -> {
+                        alert.setTitle(R.string.export_error_label);
+                        alert.setMessage(error.getMessage());
+                        alert.show();
+                    }, progressDialog::dismiss);
+        }
         return true;
     }
 
@@ -184,22 +236,22 @@ public class StoragePreferencesFragment extends PreferenceFragmentCompat {
             String dir = data.getStringExtra(DirectoryChooserActivity.RESULT_SELECTED_DIR);
 
             File path;
-            if(dir != null) {
+            if (dir != null) {
                 path = new File(dir);
             } else {
                 path = getActivity().getExternalFilesDir(null);
             }
             String message = null;
-            final Context context= getActivity().getApplicationContext();
-            if(!path.exists()) {
+            final Context context = getActivity().getApplicationContext();
+            if (!path.exists()) {
                 message = String.format(context.getString(R.string.folder_does_not_exist_error), dir);
-            } else if(!path.canRead()) {
+            } else if (!path.canRead()) {
                 message = String.format(context.getString(R.string.folder_not_readable_error), dir);
-            } else if(!path.canWrite()) {
+            } else if (!path.canWrite()) {
                 message = String.format(context.getString(R.string.folder_not_writable_error), dir);
             }
 
-            if(message == null) {
+            if (message == null) {
                 Log.d(TAG, "Setting data folder: " + dir);
                 UserPreferences.setDataFolder(dir);
                 setDataFolderText();
@@ -209,6 +261,12 @@ public class StoragePreferencesFragment extends PreferenceFragmentCompat {
                 ab.setPositiveButton(android.R.string.ok, null);
                 ab.show();
             }
+        }
+
+        if (resultCode == Activity.RESULT_OK &&
+                requestCode == CHOOSE_OPML_EXPORT_PATH) {
+            Uri uri = data.getData();
+            export(new OpmlWriter(), uri);
         }
     }
 
@@ -229,6 +287,24 @@ public class StoragePreferencesFragment extends PreferenceFragmentCompat {
         Activity activity = getActivity();
         Intent intent = new Intent(activity, DirectoryChooserActivity.class);
         activity.startActivityForResult(intent, DirectoryChooserActivity.RESULT_CODE_DIR_SELECTED);
+    }
+
+    private void openExportPathPicker() {
+        if (Build.VERSION.SDK_INT > Build.VERSION_CODES.KITKAT_WATCH) {
+            Intent intentPickAction = new Intent(Intent.ACTION_OPEN_DOCUMENT_TREE);
+            // Creates an implicit intent to launch a file manager which lets
+            // the user choose a specific directory to export to.
+            try {
+                startActivityForResult(intentPickAction, CHOOSE_OPML_EXPORT_PATH);
+                return;
+            } catch (ActivityNotFoundException e) {
+                Log.e(TAG, "No activity found. Should never happen...");
+            }
+        }
+
+        // If we are using a SDK lower than API 21 or the implicit intent failed
+        // fallback to the legacy export process
+        export(new OpmlWriter());
     }
 
     private void showChooseDataFolderDialog() {
